backend/requirements.txt
Django==5.1.*
djangorestframework==3.15.*
psycopg2-binary==2.9.*
gunicorn==22.*
python-dotenv==1.0.*
anthropic==0.40.*
google-generativeai==0.8.*
requests==2.32.*
Pillow==11.*
